discover wizards under em folder.
